\name{k2ri}
\alias{k2ri}
\alias{gene2ec}
\alias{gene2ko}
\alias{keggGene2gene}
\alias{metabolicEC}
\alias{metabolicKO}
\alias{nonMetabolicKO}
\docType{data}

\title{The variables in the environment variable k2ri of the system}
\description{
  The variables in the environment variable k2ri of the system.
}
\details{
The environment variable includes the variable \code{gene2ec}, \code{gene2ko}, \code{keggGene2gene}, \code{metabolicEC}, \code{metabolicKO}, \code{nonMetabolicKO}, etc.

}
\format{An environment variable}
\author{ Chunquan Li \email{lcqbio@yahoo.com.cn}}
\seealso{\code{\link{loadK2ri}},\code{\link{saveK2ri}},\code{\link{initializeK2ri}}}
\keyword{file}